requirements.txt
django==4.2.1
django-cors-headers==4.0.0
motor==3.1.2
python-dotenv==1.0.0
openai==0.27.7
redis==4.5.5
hiredis==2.2.3